refactor AssessmentAdapter classes and interfaces
